updated admin action in delete and  update memberships

+++ controllers/membershipFolder.js
@@ -58,12 +58,21 @@ exports.getadminFolders = (req, res) => {
     });
 };
 exports.update_folder = (req, res) => {
+  const adminId = req.params.adminId
+  const userId = req.params.userId;
+  const folderId = req.params.folderId
   membershipFolder
-    .findByIdAndUpdate(req.params.folderId, req.body)
+    .updateOne({ _id: folderId, $and: [{ userId: userId }, { adminId: adminId }] }, { $set: req.body })
     .exec((err, updateFolder) => {
       if (err) {
         res.send({ msg: "membership folder is not updated", success: false });
       } else {
+        if (updateFolder.n < 1) {
+          return res.status(401).send({
+            msg: "This is system generated folder Only admin can update",
+            success: false,
+          });
+        }
         res.send({
           msg: "Folder is update successfully",
           success: true,
@@ -73,12 +82,21 @@ exports.update_folder = (req, res) => {
 };
 
 exports.delete_folder = (req, res) => {
+  const adminId = req.params.adminId
+  const userId = req.params.userId;
+  const folderId = req.params.folderId
   membershipFolder.findOneAndRemove(
-    { _id: req.params.folderId },
+    { _id: folderId, $and: [{ userId: userId }, { adminId: adminId }] },
     (err, delFolder) => {
       if (err) {
         res.send({ msg: "Folder is not remove", success: false });
       } else {
+        if (!delFolder) {
+          return res.status(401).send({
+            msg: "This is system generated folder Only admin can update",
+            success: false,
+          });
+        }
         membership.deleteMany(
           { folderId: req.params.folderId },
           (err, delFolder) => {
